Fix OoE event var Bs being reset under specific circumstances
Normally the ooe_nonlinear patch unsets these var Bs, but if room rando
is on the rooms get cached in memory before applying the patch. Then if
those rooms are changed (such as by adding a wooden door to a room) the
old cached var B is written back to rom, undoing the what the patch did.
This just manually unsets the var Bs again for all relevant events.