Place starting portrait in a random valid location
Previously if you had starting room rando on and it started you inside a
portrait, that portrait would be placed in the vanilla location in the
castle. But that would be a problem if that room in the castle is not
used by the map rando. So instead place it in a random location in the
castle that is placed by the map rando.