Fix bias bug where enemies would have more weaknesses than resistances
If both a weakness and resistance bit got set, the randomizer would
unset the resistance bit to not cause bugs, but that lead to a heavy
bias towards weaknesses. Here we change it to be a random 50% chance
whether the weakness or resistance bit is unset. This way they should be
about even (assuming the weakness and resistance difficulty options are
set to the same value).