Enemy sprite rando: Fix sprites not being loaded correctly
If an enemy originally called LoadSpriteSingleGfx and we swapped that
call to be LoadSpriteMultiGfx, then the extra argument taken by
LoadSpriteMultiGfx would not be set correctly, causing buggy behavior.

We add a new custom function that takes the same arguments as
LoadSpriteSingleGfx, but calls LoadSpriteMultiGfx with the extra
argument being set to 1 (which is what it always should be).

Then instead of swapping LoadSpriteSingleGfx for LoadSpriteMultiGfx, we
swap it for this new custom function.