Reduce max map rando map width from 60 to 59 to fix minor visual bugs
One bug is that the map could slightly overlap the outer border of the
top screen. The other is that the map could get its rightmost tile onto
a special tile space which is supposed to be blank - if that happened
the whole screen would be covered with that tile very briefly when the
player presses up at a warp point.