Create option to skip drop chance randomization when randomizing enemy drops.
Currently, if you select the randomize drops option, the drop chances will be randomized as well to fit the slider options in difficulty tab. While in a normal randomizer run this allows you to get drops more often, it causes races or special runs like ALL souls to feel off since the rates aren't vanilla.

This fork proposes a small change to UI and code which allows for the randomization of drops without randomizing chances.

I still have to test whether this has an impact on progress drops, and there's a possible change to allow players to randomize drop rates without randomizing drops.